fix(homepage): fixes logo positioning in header
A recent change introduced a <meta> element within the header and, even though it is not visible,
this causes the logo to no longer be the first child element of its parent and therefore it gets a
margin-top applied to it. This pushed the logo downward, which is not what we wanted.