chore: migrates to latest Gravity next release
Previously the gravity-ui-web next tag pointed to a 1.x release. Now it points to a 3.x pre-release
which contains several breaking changes. This was causing website builds on the next branch to fix,
so it was temporarily "downgraded" to the latest 2.x release (same as on develop). This commit moves
the Gravity dependency back to next but also changes the SASS to work around the breaking changes.
Now everything builds cleanly.