fix: fixes layout bug on article pages
The ARTICLE element had an invisible META element as its first child, with Gravity's CSS, this
causes the next element (the article's heading in this case) to have a margin top added to it.
However, in this instance it was unintentional and caused the article headings to appear slightly
lower than headings on other pages. This was being flagged up by Gravity's debug CSS. This change
fixes the issue by moving the META element down so that it occurs after the heading. This hasn't
altered the structured meatadata in any way (I verified it in Google's structured data testing tool)