feat: uses theme and favicon colors from Gravity
We use some metadata to configure the appearance of web browser UIs and favicons for certain
browsers and OSes. Previsouly the color used was hardcoded (and based on Gravity 2.x's color
palette). This change pulls in appropriate color purposes from Gravity's particles package so that
the colors used will always be in sync with Gravity's color palette.