feat: adds favicon generation to build
So far, we had a static set of website icons in various sizes and formats. This makes it harder to
update the icons - either due to logo changes or due to color changes in Gravity. This change
therefore introduces a new Gulp task in our build that will generate all the required icon images
and sizes from a single source image.