feat: adds image optimisation gulp plugin
Adds a custom Gulp plug-in that can crop and resize source images to defined sizes (or size ranges)
in order to reduce their file sizes. Without this, the homepage currently clocks up 1.35mb in total.
With this, that gets reduced down to 475kb - about a third! (On single density screens, the figure
is even lower)