fix(a11y): ensures pa11y tests always run against localhost
Pa11y uses the sitemap.xml to get the list of URLs to check. For non-dev builds, the sitemap
contains absolute URLs to our staging and production environments, meaning pa11y checks those rather
than the local copy of the website that's been built. This is obviously not what we want and can
lead to weird build failures due to an old issue on the staging env that has been fixed locally.
This change forces pa11y to always run against localhost.